production changes

+++ backend/app/models.py
@@ -1,5 +1,4 @@
 from pydantic import BaseModel, Field
-from typing import Optional, List
 from datetime import datetime
 
 class Trip(BaseModel):
@@ -7,18 +6,18 @@ class Trip(BaseModel):
     trip_date: str = Field(default_factory=lambda: datetime.utcnow().strftime('%Y-%m-%d'))
     total_trips: int = Field(ge=0)
     working_hours: str
-    earnings_ola: float = Field(default=0.0, ge=0)
-    earnings_uber: float = Field(default=0.0, ge=0)
-    earnings_rapido: float = Field(default=0.0, ge=0)
-    gross_earnings: float = Field(ge=0)
+    earnings_ola: int = Field(default=0, ge=0)
+    earnings_uber: int = Field(default=0, ge=0)
+    earnings_rapido: int = Field(default=0, ge=0)
+    gross_earnings: int = Field(ge=0)
 
 class Expense(BaseModel):
     date: datetime = Field(default_factory=datetime.utcnow)
     expense_date: str = Field(default_factory=lambda: datetime.utcnow().strftime('%Y-%m-%d'))
-    fuel: float = Field(default=0.0, ge=0)
-    other: float = Field(default=0.0, ge=0)
+    fuel: int = Field(default=0, ge=0)
+    other: int = Field(default=0, ge=0)
 
 class Profile(BaseModel):
     name: str
     car_model: str
-    rating: float = Field(default=5.0, ge=0, le=5) 
+    rating: int = Field(default=5, ge=0, le=5)
